many: rename tests as <component>.test.<ext>
