+++ src/components/FamilyConnectionsModal/FamilyConnectionsModal.test.jsx
@@ -4,19 +4,19 @@ import { Platform, View } from 'react-native';
 import renderer from 'react-test-renderer';
 import { render, fireEvent } from 'react-native-testing-library';
 
-import RecentSearches from './index.jsx';
+import FamilyConnectionsModal from './FamilyConnectionsModal';
 
 /**********************************************************/
 
-describe('RecentSearches component', () => {
+describe('FamilyConnectionsModal component', () => {
 
     test('renders', () => {
-        const tree = renderer.create(<RecentSearches/>).toJSON();
+        const tree = renderer.create(<FamilyConnectionsModal/>).toJSON();
         expect(tree.children).toHaveLength(1);
     });
 
     test('matches snapshot', () => {
-        const tree = renderer.create(<RecentSearches/>).toJSON();
+        const tree = renderer.create(<FamilyConnectionsModal/>).toJSON();
         expect(tree).toMatchSnapshot();
     });
