+++ src/components/RecentSearches/RecentSearches.test.jsx
@@ -4,19 +4,19 @@ import { Platform, View } from 'react-native';
 import renderer from 'react-test-renderer';
 import { render, fireEvent } from 'react-native-testing-library';
 
-import FamilyConnectionsModal from './index.jsx';
+import RecentSearches from './RecentSearches';
 
 /**********************************************************/
 
-describe('FamilyConnectionsModal component', () => {
+describe('RecentSearches component', () => {
 
     test('renders', () => {
-        const tree = renderer.create(<FamilyConnectionsModal/>).toJSON();
+        const tree = renderer.create(<RecentSearches/>).toJSON();
         expect(tree.children).toHaveLength(1);
     });
 
     test('matches snapshot', () => {
-        const tree = renderer.create(<FamilyConnectionsModal/>).toJSON();
+        const tree = renderer.create(<RecentSearches/>).toJSON();
         expect(tree).toMatchSnapshot();
     });
